# docs/swagger.yaml
basePath: /
definitions:
...
  JSON:
    additionalProperties: {}
    type: object
...
  agentListItem:
    properties:
      attached_skills:
        items:
          $ref: '#/definitions/agentSkillSummary'
        type: array
      avatar_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: string
      instructions:
        type: string
      is_default:
        type: boolean
      last_memory_refreshed_at:
        type: string
      mcp_servers:
        items:
          type: integer
        type: array
      memory_refresh_error:
        type: string
      memory_refresh_status:
        type: string
      model:
        type: string
      name:
        type: string
      permissions:
        $ref: '#/definitions/JSON'
      placeholder:
        type: string
      resources:
        $ref: '#/definitions/JSON'
      sandbox:
        $ref: '#/definitions/agentSandboxSummary'
      sandbox_strategy:
        type: string
      sandbox_template_id:
        type: string
      sandbox_tools:
        items:
          type: string
        type: array
      skills:
        $ref: '#/definitions/JSON'
      status:
        type: string
      tools:
        $ref: '#/definitions/JSON'
      triggers:
        items:
          $ref: '#/definitions/agentTriggerResponse'
        type: array
      updated_at:
        type: string
      upgrade_available:
        type: boolean
    type: object
  agentMutationRequest:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      icon:
        type: string
      instructions:
        type: string
      mcp_servers:
        items:
          type: integer
        type: array
      model:
        type: string
      name:
        type: string
      permissions:
        $ref: '#/definitions/JSON'
      placeholder:
        type: string
      resources:
        $ref: '#/definitions/JSON'
      sandbox_strategy:
        type: string
      sandbox_template_id:
        type: string
      sandbox_tools:
        items:
          type: string
        type: array
      skills:
        $ref: '#/definitions/JSON'
      tools:
        $ref: '#/definitions/JSON'
    type: object
  agentMutationResponse:
    properties:
      agent:
        $ref: '#/definitions/agentListItem'
    type: object
...
  agentSandboxSummary:
    properties:
      created_at:
        type: string
      error_message:
        type: string
      external_id:
        type: string
      id:
        type: string
      last_active_at:
        type: string
      status:
        type: string
    type: object
...
  agentSkillSummary:
    properties:
      description:
        type: string
      id:
        type: string
      locked:
        type: boolean
      name:
        type: string
      required:
        type: boolean
      source_type:
        type: string
    type: object
  agentTriggerResponse:
    properties:
      conditions: {}
      connection_id:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      instructions:
        type: string
      provider:
        type: string
      secret_set:
        description: |-
          SecretSet indicates whether an HTTP trigger has a shared secret configured.
          True when the trigger requires auth on incoming requests. The secret value
          is never returned.
        type: boolean
      trigger_keys:
        items:
          type: string
        type: array
      trigger_type:
        type: string
    type: object
...
  errorResponse:
    properties:
      error:
        type: string
    type: object
...
host: api.dev.usehivy.com
info:
  contact: {}
  description: Proxy runtime for LLM API credentials.
  title: Hivy API
  version: "1.0"
paths:
...
  /v1/agents/{id}:
    delete:
      description: Archives an agent when it is not the default Hivy agent and has
        no active sessions.
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agentMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Archive an agent
      tags:
      - agents
    get:
      description: |-
        Returns one agent in the org with skills (metadata only — no bundle content),
        triggers, and the latest sandbox row.
      parameters:
      - description: Agent agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agentListItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Get an AI agent
      tags:
      - agents
    patch:
      consumes:
      - application/json
      description: Updates a user-managed agent definition. The default Hivy agent
        cannot be renamed or moved away from always_on.
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Agent update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/agentMutationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agentMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorResponse'
      security:
      - BearerAuth: []
      summary: Update an agent
      tags:
      - agents
...
  /v1/assets/preview:
    get:
      description: Redirects to a short-lived signed URL for an object in the public
        assets bucket.
      parameters:
      - description: Public assets S3 key
        in: query
        name: path
        required: true
        type: string
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/errorResponse'
      summary: Preview public asset
      tags:
      - assets
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Bearer token (JWT or API key).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"